requirements.txt
Flask>=2.2.5
Flask-SQLAlchemy==3.1.1
APScheduler==3.10.4
email-validator==2.1.0.post1
google-auth==2.22.0
google-auth-oauthlib==1.0.0
oauth2client==4.1.3
google-api-python-client==2.95.0
requests==2.31.0
google-auth-httplib2==0.1.0
httplib2==0.22.0
python-dotenv==0.19.0 